optimization-减少冗余的跨链地址列表分享
