+++ module/nuls-network/src/main/java/io/nuls/network/manager/handler/message/GetAddrMessageHandler.java
@@ -25,24 +25,18 @@
 
 package io.nuls.network.manager.handler.message;
 
-import io.nuls.core.core.ioc.SpringLiteContext;
-import io.nuls.core.log.Log;
-import io.nuls.network.cfg.NetworkConfig;
 import io.nuls.network.manager.MessageFactory;
 import io.nuls.network.manager.MessageManager;
 import io.nuls.network.manager.NodeGroupManager;
 import io.nuls.network.manager.handler.base.BaseMessageHandler;
 import io.nuls.network.model.NetworkEventResult;
 import io.nuls.network.model.Node;
-import io.nuls.network.model.NodeGroup;
 import io.nuls.network.model.dto.IpAddressShare;
 import io.nuls.network.model.message.AddrMessage;
 import io.nuls.network.model.message.base.BaseMessage;
 import io.nuls.network.model.message.body.GetAddrMessageBody;
 import io.nuls.network.utils.LoggerUtil;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 
 /**
@@ -79,7 +73,7 @@ public class GetAddrMessageHandler extends BaseMessageHandler {
         LoggerUtil.logger(chainId).info("GetAddrMessageHandler Recieve:" + (node.isServer() ? "Server" : "Client") + ":" + node.getIp() + ":" + node.getRemotePort() + "==CMD=" + message.getHeader().getCommandStr());
         GetAddrMessageBody getAddrMessageBody = (GetAddrMessageBody) message.getMsgBody();
         //发送addr消息
-        List<IpAddressShare> ipAddresses = getAvailableNodes(node, getAddrMessageBody.getChainId(), (getAddrMessageBody.getIsCrossAddress() == (byte) 1 ? true : false));
+        List<IpAddressShare> ipAddresses = NodeGroupManager.getInstance().getAvailableShareNodes(node, getAddrMessageBody.getChainId(), (getAddrMessageBody.getIsCrossAddress() == (byte) 1 ? true : false));
         AddrMessage addressMessage = MessageFactory.getInstance().buildAddrMessage(ipAddresses, message.getHeader().getMagicNumber(), getAddrMessageBody.getChainId(), getAddrMessageBody.getIsCrossAddress());
         if (0 == addressMessage.getMsgBody().getIpAddressList().size()) {
             LoggerUtil.logger(chainId).info("No Address");
@@ -94,74 +88,6 @@ public class GetAddrMessageHandler extends BaseMessageHandler {
     public NetworkEventResult send(BaseMessage message, Node node, boolean asyn) {
         LoggerUtil.logger(node.getNodeGroup().getChainId()).info("GetAddrMessageHandler Send:" + (node.isServer() ? "Server" : "Client") + ":" + node.getIp() + ":" + node.getRemotePort() + "==CMD=" + message.getHeader().getCommandStr());
         return super.send(message, node, asyn);
-    }
-
-    /**
-     * 获取可分享的节点信息
-     *
-     * @param node
-     * @return
-     */
-    private List<IpAddressShare> getAvailableNodes(Node node, int getChainId, boolean isCrossAddress) {
-        NetworkConfig networkConfig = SpringLiteContext.getBean(NetworkConfig.class);
-
-        List<IpAddressShare> addressList = new ArrayList<>();
-        List nodesList = new ArrayList();
-        if (node.isCrossConnect()) {
-            //跨链节点的请求
-            //取本地网络地址去支持跨链连接,跨链的请求地址取的都是对方的本地网络IP
-            if (networkConfig.isMoonNode()) {
-                //是主网节点，回复
-                nodesList.addAll(NodeGroupManager.getInstance().getMoonMainNet().getLocalShareToCrossCanConnectNodes().values());
-            } else {
-                nodesList.addAll(node.getNodeGroup().getLocalShareToCrossCanConnectNodes().values());
-            }
-        } else {
-            //非跨链节点的请求,分2种，一种是获取外界网络的跨链地址，一种是自身网络地址
-            NodeGroup nodeGroup = NodeGroupManager.getInstance().getNodeGroupByChainId(getChainId);
-            if (isCrossAddress) {
-                //主网 跨链网络组
-                nodesList.addAll(nodeGroup.getCrossNodeContainer().getAllCanShareNodes().values());
-            } else {
-                nodesList.addAll(nodeGroup.getLocalNetNodeContainer().getAllCanShareNodes().values());
-            }
-        }
-
-        addAddress(nodesList, addressList, node.getIp(), node.isCrossConnect());
-        return addressList;
-
-    }
-
-    private void addAddress(Collection<Node> nodes, List<IpAddressShare> list, String fromIp, boolean isCross) {
-        for (Node peer : nodes) {
-            /*
-             * 排除自身连接信息，比如组网A=====B，A向B请求地址，B给的地址列表需排除A地址。
-             * Exclude self-connection information, such as networking A=====B,
-             * A requests an address from B, and the address list given by B excludes the A address.
-             */
-            if (peer.getIp().equals(fromIp)) {
-                continue;
-            }
-            /*
-             * 只有主动连接的节点地址才可使用。
-             * Only active node addresses are available for use.
-             */
-            if (Node.OUT == peer.getType()) {
-                try {
-                    int port = peer.getRemotePort();
-                    int crossPort = peer.getRemoteCrossPort();
-                    if (isCross) {
-                        if (0 == crossPort) {
-                            continue;
-                        }
-                    }
-                    IpAddressShare ipAddress = new IpAddressShare(peer.getIp(), port, crossPort);
-                    list.add(ipAddress);
-                } catch (Exception e) {
-                    Log.error(e);
-                }
-            }
-        }
     }
 
 }
