+++ module/nuls-network/src/main/java/io/nuls/network/manager/handler/message/VersionMessageHandler.java
@@ -47,7 +47,6 @@ import io.nuls.network.rpc.call.BlockRpcService;
 import io.nuls.network.rpc.call.impl.BlockRpcServiceImpl;
 import io.nuls.network.utils.LoggerUtil;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -222,14 +221,6 @@ public class VersionMessageHandler extends BaseMessageHandler {
             //向种子节点请求地址
             MessageManager.getInstance().sendGetAddressMessage(node, false, false, true);
         }
-        List<NodeGroup> nodeGroupList = NodeGroupManager.getInstance().getNodeGroups();
-        nodeGroupList.forEach(nodeGroup -> {
-            //client像主网节点请求跨链地址列表
-            if (nodeGroup.isMoonCrossGroup()) {
-                MessageManager.getInstance().sendGetCrossAddressMessage(nodeGroupManager.getMoonMainNet(), nodeGroup, false, true, true);
-            }
-        });
-
     }
 
     /**
